separate Tex and Dds logic
