+++ Arrowgene.Ddon.Client/Resource/Texture/Tex/TexLayer.cs
@@ -1,0 +1,8 @@
+﻿namespace Arrowgene.Ddon.Client.Resource.Texture.Tex;
+
+public struct TexLayer
+{
+    public uint Offset;
+    public uint Size;
+    public byte[] Data;
+}
