+++ Arrowgene.Ddon.Client/Resource/Texture/Tex/TexPixelFormat.cs
@@ -1,0 +1,60 @@
+﻿namespace Arrowgene.Ddon.Client.Resource.Texture.Tex;
+
+public enum TexPixelFormat
+{
+    FORMAT_UNKNOWN = 0,
+    FORMAT_R32G32B32A32_FLOAT = 1,
+    FORMAT_R16G16B16A16_FLOAT = 2,
+    FORMAT_R16G16B16A16_UNORM = 3,
+    FORMAT_R16G16B16A16_SNORM = 4,
+    FORMAT_R32G32_FLOAT = 5,
+    FORMAT_R10G10B10A2_UNORM = 6,
+    FORMAT_R8G8B8A8_UNORM = 7,
+    FORMAT_R8G8B8A8_SNORM = 8,
+    FORMAT_R8G8B8A8_UNORM_SRGB = 9,
+    FORMAT_B4G4R4A4_UNORM = 0x0A,
+    FORMAT_R16G16_FLOAT = 0x0B,
+    FORMAT_R16G16_UNORM = 0x0C,
+    FORMAT_R16G16_SNORM = 0x0D,
+    FORMAT_R32_FLOAT = 0x0E,
+    FORMAT_D24_UNORM_S8_UINT = 0x0F,
+    FORMAT_R16_FLOAT = 0x10,
+    FORMAT_R16_UNORM = 0x11,
+    FORMAT_A8_UNORM = 0x12,
+    FORMAT_BC1_UNORM = 0x13,
+    FORMAT_BC1_UNORM_SRGB = 0x14,
+    FORMAT_BC2_UNORM = 0x15,
+    FORMAT_BC2_UNORM_SRGB = 0x16,
+    FORMAT_BC3_UNORM = 0x17,
+    FORMAT_BC3_UNORM_SRGB = 0x18,
+    FORMAT_BCX_GRAYSCALE = 0x19,
+    FORMAT_BCX_ALPHA = 0x1A,
+    FORMAT_BC5_SNORM = 0x1B,
+    FORMAT_B5G6R5_UNORM = 0x1C,
+    FORMAT_B5G5R5A1_UNORM = 0x1D,
+    FORMAT_BCX_NM1 = 0x1E,
+    FORMAT_BCX_NM2 = 0x1F,
+    FORMAT_BCX_RGBI = 0x20,
+    FORMAT_BCX_RGBY = 0x21,
+    FORMAT_B8G8R8X8_UNORM = 0x22,
+    FORMAT_BCX_RGBI_SRGB = 0x23,
+    FORMAT_BCX_RGBY_SRGB = 0x24,
+    FORMAT_BCX_NH = 0x25,
+    FORMAT_R11G11B10_FLOAT = 0x26,
+    FORMAT_B8G8R8A8_UNORM = 0x27,
+    FORMAT_B8G8R8A8_UNORM_SRGB = 0x28,
+    FORMAT_BCX_RGBNL = 0x29,
+    FORMAT_BCX_YCCA = 0x2A,
+    FORMAT_BCX_YCCA_SRGB = 0x2B,
+    FORMAT_R8_UNORM = 0x2C,
+    FORMAT_B8G8R8A8_UNORM_LE = 0x2D,
+    FORMAT_B10G10R10A2_UNORM_LE = 0x2E,
+    FORMAT_BCX_SRGBA = 0x2F,
+    FORMAT_BC7_UNORM = 0x30,
+    FORMAT_BC7_UNORM_SRGB = 0x31,
+    FORMAT_SE5M9M9M9 = 0x32,
+    FORMAT_R10G10B10A2_FLOAT = 0x33,
+    FORMAT_YVU420P2_CSC1 = 0x34,
+    FORMAT_R8A8_UNORM = 0x35,
+    FORMAT_A8_UNORM_WHITE = 0x36
+}
