+++ Arrowgene.Ddon.Client/Resource/Texture/Tex/TexSphericalHarmonicsVector.cs
@@ -1,0 +1,8 @@
+﻿namespace Arrowgene.Ddon.Client.Resource.Texture.Tex;
+
+public struct TexSphericalHarmonicsVector
+{
+    public float R;
+    public float G;
+    public float B;
+}
